requirements.txt
Django==5.0.10
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Pillow==10.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
openpyxl==3.1.5